+++ backend/hqlib/__init__.py
@@ -15,4 +15,4 @@ limitations under the License.
 """
 
 NAME = "HQ"
-VERSION = "2.83.2"
+VERSION = "2.84.0"